+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add `projects create` command to create new projects
+
 ### Fixed
 
 ## [0.4.0] - 2017-08-31